+++ website/common/script/ops/buy/buyQuestGold.js
